웹에서 로그인 세션을 직접 발급
