+++ .env
@@ -1,0 +1,9 @@
+DATABASE_URL=
+PUBLIC_API_ORIGIN=
+PUBLIC_OIDC_CLIENT_ID=
+PUBLIC_OIDC_ISSUER=
+PUBLIC_ORIGIN=
+OIDC_CLIENT_ID=
+OIDC_CLIENT_SECRET=
+OIDC_ISSUER=
+OIDC_REDIRECT_URI=
